stash 2 of tryna switch branches

+++ src/src/Karma/Category.java
@@ -5,15 +5,17 @@ import x3mara.*;
 import Jasmin.*;
 
 import java.util.Arrays;
+import java.util.InputMismatchException;
 import java.util.Objects;
 import java.util.Scanner;
+import java.util.ArrayList;
+
 
 public class Category implements HasID {
     private final String catID; //store the ID of each category
     private String catName; // name of the category
     public static int totCats = 0; // stores total number of categories;
-    private Event[] events = new Event[100];  // stores events under each category;
-    private int numEvents = 0; // stores number of events in each category
+    private ArrayList<Event> events = new ArrayList<>();// stores events under each category;
 
 
     Scanner input = new Scanner(System.in);
@@ -22,7 +24,6 @@ public class Category implements HasID {
     // no-arg constructor
     public Category() {
         this.catID = "C" + System.nanoTime();
-        totCats++;
     }
 
     // Arg constructor
@@ -40,8 +41,8 @@ public class Category implements HasID {
         return catName;
     }
 
-    public Event[] getEvents() {
-        return Arrays.copyOf(events, numEvents); // to ignore the null values
+    public ArrayList<Event> getEvents() {
+        return events; // to ignore the null values
     }
 
 
@@ -49,27 +50,74 @@ public class Category implements HasID {
         this.catName = catName;
     }
 
-    public void setEvents(Event[] events) {
+    public void setEvents(ArrayList<Event> events) {
         this.events = events;
     }
 
-    public final void ValidateCatAccess(User obj){
+    public static void ValidateCatAccess(User obj){
         if (!(obj instanceof Admin)) {
             throw new AccessDenied("You do not have permission to use this method." +
                     " \n Only Admins are allowed to create categories");
         }
     }
-    public void addCatToDatabase(User obj) {
+    public static void addCatToDatabase(User obj) {
         ValidateCatAccess(obj);
-        Database.create(this);
+        System.out.println("Enter the name of the Category");
+        Scanner input = new Scanner(System.in);
+        String name = input.next();
+        Database.create(new Category(name));
     }
     public void updateCatInDatabase(User obj){
         ValidateCatAccess(obj);
-        Database.update(this);
+        System.out.println("What would you like to change?");
+        String[] options1 = new String[]{"Category Name","Exit"};
+        Scanner in = new Scanner(System.in);
+        System.out.println("Choose an option:");
+        for(int i=0;i<options1.length;i++){
+            System.out.println((i+1) + ") " + options1[i]);
+        }
+        int choice = in.nextInt();
+        if(choice < 1 || choice > options1.length){
+            throw new InputMismatchException("Input should be inbounds.");
+        }
+        switch (choice){
+            case 1:
+                System.out.println("Original Data : \nCategory Name : " + getCatName());
+                System.out.println("Choose an option:\n1. keep\n2. change");
+                int choice2 = in.nextInt();
+                if(choice2 < 1 || choice2 >2){
+                    throw new InputMismatchException("Input should be inbounds.");
+                }
+                switch(choice2){
+                    case 1 :
+                        System.out.println("change has been canceled");
+                        break;
+
+                    case 2:
+                        System.out.println("Enter a new Name");
+                        setCatName(in.next());
+                        break;
+                }
+
+            case 2:
+                break;
+        }
+
     }
     public void deleteCatFromDatabase(User obj){
         ValidateCatAccess(obj);
-        Database.delete(this);
+        System.out.println("Are you sure you want to delete this Category? : 1. yes, 2. No");
+        int choice = input.nextInt();
+        if(choice == 1){
+            Database.delete(this);
+        }
+        else if (choice == 2){
+            System.out.println("Deletion canceled");
+        }
+        else{
+            throw new InputMismatchException("Input should be inbounds.");
+        }
+
     }
 
     public static void listAllCategories(){
@@ -77,14 +125,17 @@ public class Category implements HasID {
     }
 
     public void addEvent(Event event) {
-        if (numEvents > 100) { // to avoid exceeding the limit of the array
-            throw new ExceedLimit("You have reached the limit of events for " +
-                    "a category");
-        }
-        else {
-            events[numEvents++] = event;
+        if (events.size() >= 100) {
+            throw new ExceedLimit("You have reached the limit of events for a category");
+        } else {
+            events.add(event);
         }
     }
+
+    public void deleteEventFromCat(Event obj) {
+        events.remove(obj);
+    }
+
 
     @Override
     public String toString() {
